pyproject.toml
[project]
name = "jb-peak-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter-tabnine>=1.2.3",
    "matplotlib>=3.10.3",
    "nbclassic>=1.3.1",
    "notebook>=7.4.3",
    "numpy>=2.3.0",
    "pillow>=11.2.1",
    "pip>=25.1.1",
    "scipy>=1.15.3",
]
